- Timer will be removed from timer-array at the same time it has zero-counted, so   there is no gap between timer elapsed and timer removed. - Timer is retriggerable by now.

+++ sw/src/shared/clock.h
@@ -24,7 +24,7 @@
 #define CLOCK_TICKS_PER_SECOND  100
 
 //! Convert from milliseconds into ticks
-#define CLOCK_MS_2_TICKS(ms)    (uint8_t)((uint32_t)ms*CLOCK_TICKS_PER_SECOND/1000)
+#define CLOCK_MS_2_TICKS(ms)    (uint16_t)((uint32_t)ms*CLOCK_TICKS_PER_SECOND/1000)
 
 // --- Type definitions --------------------------------------------------------
